Port remaining writable-stream.js tests to wpt
This is an initial port of the 3 remaining tests from
test/writable-stream.js to web-platform-tests.

Minimal changes have been made. A followup commit will change the new
tests from async_test to promise_test and use a more idiomatic style.
